fix: move firebase logic
