+++ src/hooks/useRankingData.js
@@ -1,15 +1,12 @@
 import { useEffect, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { collection, getDocs, query, orderBy, where } from 'firebase/firestore';
 
-import { GAME_STATUS, RANKING_VALUES } from 'constants/types';
-import { DAILY_RESULTS, USERS, USERS_STATISTICS } from 'firebase/collections';
-import firebaseData from 'firebase/firebase';
+import { RANKING_VALUES } from 'constants/types';
+import { getAllTimeRankingData, getTodaysResults } from 'firebase/ranking';
+import { getUsers } from 'firebase/users';
 import { getTodaysDate } from 'utils/helpers';
 
 import useAuth from './useAuth';
-
-const { firebaseDb } = firebaseData;
 
 const useRankingData = () => {
   const {
@@ -29,87 +26,22 @@ const useRankingData = () => {
   useEffect(() => {
     (async function () {
       setLoading(true);
-      const q = query(collection(firebaseDb, USERS));
-      const docs = await getDocs(q);
-      const results = {};
-      docs.forEach(doc => {
-        const user = doc.data();
-        results[user.email] = user;
-      });
-      setUsers(results);
+      const { users: usersResults } = await getUsers();
+      setUsers(usersResults);
     })();
   }, []);
 
   useEffect(() => {
     (async function () {
-      const q = query(
-        collection(firebaseDb, DAILY_RESULTS),
-        where('date', '==', today),
-        where('status', '!=', GAME_STATUS.playing),
-        orderBy('status', 'desc'),
-        orderBy('attempts'),
-        orderBy('solveTime'),
-        orderBy('user.name')
-      );
-      const docs = await getDocs(q);
-
-      const results = [];
-      let position = 0;
-      let currentAttempts = 0;
-      let currentStatus = GAME_STATUS.won;
-      let currentSolveTime = 0;
-      docs.forEach(doc => {
-        const { attempts, solveTime, status, ...restDailyResults } = doc.data();
-        if (
-          currentAttempts !== attempts ||
-          solveTime !== currentSolveTime ||
-          currentStatus !== status
-        ) {
-          currentAttempts = attempts;
-          currentSolveTime = solveTime;
-          currentStatus = status;
-          position += 1;
-        }
-        results.push({
-          attempts,
-          solveTime,
-          status,
-          ...restDailyResults,
-          position,
-        });
-      });
-      setDailyResults(results);
+      const { todaysResults } = await getTodaysResults(today);
+      setDailyResults(todaysResults);
     })();
   }, [today]);
 
   useEffect(() => {
     (async function () {
-      const q = query(collection(firebaseDb, USERS_STATISTICS), orderBy('currentStreak', 'desc'));
-      const docs = await getDocs(q);
-      const results = [];
-      let position = 0;
-      let currentStreakValue = Number.MAX_SAFE_INTEGER;
-      docs.forEach(async doc => {
-        const { id: userEmail } = doc;
-        const { currentStreak, lastDatePlayed, totalGames, ...restStatistics } = doc.data();
-        if (!!totalGames) {
-          if (currentStreak !== currentStreakValue) {
-            currentStreakValue = currentStreak;
-            position += 1;
-          }
-          results.push({
-            ...restStatistics,
-            currentStreak,
-            lastDatePlayed,
-            rightText: RANKING_VALUES[0].getRightText({ currentStreak, isFirst: position === 1 }),
-            suffix: RANKING_VALUES[0].getSuffix({ lastDatePlayed }),
-            totalGames,
-            position,
-            user: users[userEmail] || { email: userEmail, name: userEmail.split('@')[0] },
-          });
-        }
-      });
-      setRankingData(results);
+      const { allTimeRankingData } = await getAllTimeRankingData(users);
+      setRankingData(allTimeRankingData);
       setLoading(false);
     })();
   }, [users]);
